chore(rules): reorganise research rules
