+++ cds_migrator_kit/rdm/records/transform/models/research.py
@@ -33,8 +33,6 @@ class ResearchModel(CdsOverdo):
         "520__9",  # provenance of the description
         "520__h",  # provenance of the description
         "852__c",  # holdings will be taken separately
-        "583__c",  # Action note date e.g 1416779
-        "583__z",  # Action note e.g 1416779
         "852__h",
         "037__c",  # arxiv subject
         "100__m",  # email of contributor
