+++ setup.cfg
@@ -93,13 +93,13 @@ cds_migrator_kit.migrator.rules.base =
 cds_migrator_kit.migrator.rdm.rules.base =
     base = cds_migrator_kit.rdm.records.transform.xml_processing.rules.base
 cds_migrator_kit.migrator.rdm.rules.publication =
-    publication = cds_migrator_kit.rdm.records.transform.xml_processing.rules.publications
+    publication = cds_migrator_kit.rdm.records.transform.xml_processing.rules.research
 cds_migrator_kit.migrator.rules.affiliations =
    affiliations = cds_migrator_kit.rdm.affiliations.xml_processing.rules.affiliations
 cds_migrator_kit.migrator.rules.ssn =
     base = cds_migrator_kit.transform.xml_processing.rules.base
     base_records = cds_migrator_kit.rdm.records.transform.xml_processing.rules.base
-    publication = cds_migrator_kit.rdm.records.transform.xml_processing.rules.publications
+    publication = cds_migrator_kit.rdm.records.transform.xml_processing.rules.research
     summer_students = cds_migrator_kit.rdm.records.transform.xml_processing.rules.summer_student_report
 cds_migrator_kit.migrator.rdm.rules.research =
     base = cds_migrator_kit.transform.xml_processing.rules.base
@@ -108,7 +108,7 @@ cds_migrator_kit.migrator.rdm.rules.research =
 cds_migrator_kit.migrator.rules.thesis =
     base = cds_migrator_kit.transform.xml_processing.rules.base
     base_records = cds_migrator_kit.rdm.records.transform.xml_processing.rules.base
-    publication = cds_migrator_kit.rdm.records.transform.xml_processing.rules.publications
+    publication = cds_migrator_kit.rdm.records.transform.xml_processing.rules.research
     thesis = cds_migrator_kit.rdm.records.transform.xml_processing.rules.thesis
 cds_migrator_kit.migrator.rules.mous =
     base = cds_migrator_kit.transform.xml_processing.rules.base
@@ -120,7 +120,7 @@ cds_migrator_kit.migrator.rules.bulletin_issue =
 cds_migrator_kit.migrator.rules.yellow_report =
     base = cds_migrator_kit.transform.xml_processing.rules.base
     base_records = cds_migrator_kit.rdm.records.transform.xml_processing.rules.base
-    publication = cds_migrator_kit.rdm.records.transform.xml_processing.rules.publications
+    publication = cds_migrator_kit.rdm.records.transform.xml_processing.rules.research
     yellow_rep = cds_migrator_kit.rdm.records.transform.xml_processing.rules.yellow_report
 cds_migrator_kit.migrator.rules.courier =
     base = cds_migrator_kit.transform.xml_processing.rules.base
@@ -130,17 +130,17 @@ cds_migrator_kit.migrator.rules.courier =
 ;cds_migrator_kit.migrator.rules.beams =
 ;    base = cds_migrator_kit.transform.xml_processing.rules.base
 ;    base_records = cds_migrator_kit.rdm.records.transform.xml_processing.rules.base
-;    publications = cds_migrator_kit.rdm.records.transform.xml_processing.rules.publications
+;    publications = cds_migrator_kit.rdm.records.transform.xml_processing.rules.research
 ;    beams = cds_migrator_kit.rdm.records.transform.xml_processing.rules.beams
 cds_migrator_kit.migrator.rules.books =
     base = cds_migrator_kit.transform.xml_processing.rules.base
     base_records = cds_migrator_kit.rdm.records.transform.xml_processing.rules.base
-    publication = cds_migrator_kit.rdm.records.transform.xml_processing.rules.publications
+    publication = cds_migrator_kit.rdm.records.transform.xml_processing.rules.research
     books = cds_migrator_kit.rdm.records.transform.xml_processing.rules.books
 cds_migrator_kit.migrator.rules.it =
     base = cds_migrator_kit.transform.xml_processing.rules.base
     base_records = cds_migrator_kit.rdm.records.transform.xml_processing.rules.base
-    publications = cds_migrator_kit.rdm.records.transform.xml_processing.rules.publications
+    publications = cds_migrator_kit.rdm.records.transform.xml_processing.rules.research
     it = cds_migrator_kit.rdm.records.transform.xml_processing.rules.it
 cds_migrator_kit.migrator.rules.it_meetings =
     base = cds_migrator_kit.transform.xml_processing.rules.base
@@ -157,17 +157,17 @@ cds_migrator_kit.migrator.rules.sy =
 cds_migrator_kit.migrator.rules.te =
     base = cds_migrator_kit.transform.xml_processing.rules.base
     base_records = cds_migrator_kit.rdm.records.transform.xml_processing.rules.base
-    publication = cds_migrator_kit.rdm.records.transform.xml_processing.rules.publications
+    publication = cds_migrator_kit.rdm.records.transform.xml_processing.rules.research
     te = cds_migrator_kit.rdm.records.transform.xml_processing.rules.te
 cds_migrator_kit.migrator.rules.en =
     base = cds_migrator_kit.transform.xml_processing.rules.base
     base_records = cds_migrator_kit.rdm.records.transform.xml_processing.rules.base
-    publication = cds_migrator_kit.rdm.records.transform.xml_processing.rules.publications
+    publication = cds_migrator_kit.rdm.records.transform.xml_processing.rules.research
     en = cds_migrator_kit.rdm.records.transform.xml_processing.rules.en
 cds_migrator_kit.migrator.rules.annual_rep =
     base = cds_migrator_kit.transform.xml_processing.rules.base
     base_records = cds_migrator_kit.rdm.records.transform.xml_processing.rules.base
-    publication = cds_migrator_kit.rdm.records.transform.xml_processing.rules.publications
+    publication = cds_migrator_kit.rdm.records.transform.xml_processing.rules.research
     annual_rep = cds_migrator_kit.rdm.records.transform.xml_processing.rules.annual_rep
 cds_migrator_kit.migrator.rules.fap =
     base = cds_migrator_kit.transform.xml_processing.rules.base
